# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto - OpenAPI 3.0
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! \
    Directly from your PC, you can upload your photos, and they will be visible to everyone who is following you.
    
    Functional design specification:
    
    - The login is performed by specifying the username and it returns a crafted userID(token).
    - Once in, the user sees on the web app the stream of photos of the users that he follows ordered in reverse chronological order.
    - Each photo has information about the date-time it was uploaded, the number of likes and the number of comments.
    - Users can place a like to photos from other users(not from themselves);
    - Users can remove likes from photos of other users;
    - Users can add comments to any photo and can edit comments to any photo(even for themselves);
    - Only the author can remove comments from a photo;
    - Users have their own personal profiles.
    
    When you access to the personal profile of a user you can see:
    - user’s photos (in reverse chronological order), 
    - their followers,
    - their following, 
    - how many photos have been uploaded.
    
    User action:
    - change his/her own username,
    - follow/unfollow other users,
    - upload photos, 
    - remove photos: also remove likes and comments.
    - Users can ban other users.
...
    Alice can decide to remove the ban at any moment.
    
...
    
...
    
...
servers:
...
components:
  parameters:
    user_id:
      description: User whose profile belongs to. 
      name: user_id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/UserID" }
    photo_id:
      description: ID of the picture considered
      name: photo_id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/PhotoID" }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Photo:
      description: Attributes of a photo
      type: object
      properties:
        photoId: {$ref: "#/components/schemas/PhotoID"}
        timestamp:
          description: Date and time of creation following RFC3339
          type: string
          format: date-time
          example: 2022-11-04T16:35:22-04:00
          minLength: 20
          maxLength: 25
        likes_count:
          description: Number of likes
          type: integer
          example: 169
        comments_count:
          description: Number of comments
          type: integer
          example: 15
        image:
          description: |
            URL of the image just uploaded.
            Accepting only ID attached with .png/.jpg/.jpeg extensions.
          type: string
          pattern: '^[a-zA-Z0-9-_]*\.(png|jpg|jpeg)$'
          example: "0a25d5f2-3b07-492e-ab87-406c4e7422b8.png"
          minLength: 0
          maxLength: 150
        caption:
          description: A written description or explanation about a photo to provide more context
          type: string
          pattern: '^[a-zA-Z0-9,._:;?!\x27\- ]*$'
          example: "Vodka may not be the answer but it's worth a shot."
          minLength: 0
          maxLength: 100
        username: {$ref: "#/components/schemas/Username"}
...
    UserID:
      description: |
        The simplified login tell us that, since there is no password, the username must be unique among the application. 
        Logging in with an existing name will grant access to that user information. 
        Hence it would make sense to use the username as ID since we don't have to represent real people but virtual people, namely usernames, by simplifying in this way the ReST OpenAPI. 
        However, since it's not applicable in a real life scenario, I will use a userID that's different from the username. Thus, I'm choosing 20 chars long strings because they can store several billions of information(64^20, since the alphabet is "[a-z][A-Z][0-9]-_", 26+26+10+2=64), they are nicer to read then integers since they are shorter and almost equally fast to compute.
      type: string
      pattern: '^[a-zA-Z0-9-_]*$'
      minLength: 1
      maxLength: 20
      example: "U3dhZ2dlciByb2Nrcw"
      readOnly: true
    PhotoID:
      description: Unique photo identifier
      type: string
      pattern: '^[a-zA-Z0-9-_]*$'
      minLength: 1
      maxLength: 20
      example: "CksvyZXtXRE"
      readOnly: true
...
    Username:
      description: It represents the name of the user in WASA Photo
      type: string
      example: LeoMessi88
      pattern: '^[a-zA-Z0-9-_]*$'
      minLength: 3
      maxLength: 16
  responses:
    BadRequest: 
        description: |-
          The request was not compliant with the documentation
          (eg. missing fields, etc).
    InternalServerError: 
      description: |-
        The server encountered an internal error.
        Further info in server logs.
    UnauthorizedError:
      description: Access token is missing or invalid.
paths:
...
  /users/{user_id}/photos/:
    parameters:
      - $ref: "#/components/parameters/user_id"
    
    get:
      tags:
       - Profile
      summary: >
        Retrieve collection of photos
      description: Retrieve collection of photos resources of a certain user
      operationId: getUserPhotos
      parameters:
        - in: query
          name: sort_by
          schema:
            description: Allowing sorting by date
            type: string
            format: date-time
            example: 2022-11-04T16:35:22-04:00
            minLength: 20
            maxLength: 25
          description: Sorting in [reverse] chronological order by date
          explode: true
        - in: query
          name: order
          schema:
            description: Sorting order
            type: string
            enum:
              - asc
              - desc
            minLength: 3
            maxLength: 4
          explode: true
              
      responses: 
        "200":
          description: List of photos ordered in reverse chronological order
          content:
            application/json: 
              schema:
                description: |
                  Represents the photos that we can see in the profile of a user. Users can see photos in batches of 50.
                type: array
                minItems: 0
                maxItems: 100
                items: { $ref: "#/components/schemas/Photo" } 
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }
      security:
        - bearerAuth: []
        
    post:
      tags:
        - Profile
      summary: Upload a photo
      description: Upload a photo on the profile of a specific user.
      operationId: uploadPhoto
      requestBody:
        content:
            application/json:
              schema: { $ref: "#/components/schemas/Photo" }
      responses: 
        "201":
          description: Photo uploaded successfully 
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Photo" } 
        "400": { $ref: "#/components/responses/BadRequest" }
        '401': {$ref: '#/components/responses/UnauthorizedError'}
        "500": { $ref: "#/components/responses/InternalServerError" }
      security:
        - bearerAuth: []
...
  /photos/{photo_id}/likes/{like_id}:
    parameters:
      - $ref: "#/components/parameters/photo_id"
      - name: like_id
        in: path
        required: true
        description: User who is liking the photo
        schema:
          { $ref: "#/components/schemas/UserID" }
    
    put:
      tags:
        - Likes
      summary: Like a photo
      description: Add a like on a photo. I'm using PUT for its idempotency.
      operationId: likePhoto
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserID'
      responses:
        "201":
          description: Like created.
          content:
            application/json: 
              schema:
                { $ref: "#/components/schemas/UserID" }
        "204":
          description: No content. Like already inserted.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        "409": { description: An user can't like its own photos. }
      security:
        - bearerAuth: []


    delete:
      tags:
        - Likes
      summary: Unlike a photo
      description: Remove a like from a photo
      operationId: unlikePhoto
      responses:
        "204":
          description: No content. Like removed.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        "409": { description: An user can't unlike its own photos. }
      security:
        - bearerAuth: []
...